requirements.txt
flask==3.0.*
gunicorn==22.0.*
pycryptodome==3.20.*